0527 많은걸 했습니다

+++ src/main/java/com/team2/docgram/dto/HashtagTableDto.java
@@ -9,6 +9,6 @@ import lombok.ToString;
 @ToString
 public class HashtagTableDto {
 	
-	private Long pk;
+	private Integer pk;
 	private String hashtagList;
 }
